Fixed tests for grids.
