+++ tests/context/test_model_context.py
@@ -17,4 +17,4 @@ def test_context_files(context):
 
 
 def test_context_location_sex(context):
-    assert str(context.db_file(1, 3, make=False)).endswith('cascade_dir/data/0/dbs/1/3/dismod.db')
+    assert str(context.db_file(1, 3)).endswith('cascade_dir/data/0/dbs/1/3/dismod.db')
